Refactor CSV helpers into separate module
